Little one pager I did after leveling up my HTML and CSS. Hope you like it. Open for comments
This is my 2nd website. I used only HTML and CSS basic elements as you will see. In future I hope to add more. Maybe a video part, and some interactivity, as well as navigation, and more pages to create a bigger project. Hope you like it, leave a comment. :)